Time: 0 ms (100%), Space: 21.8 MB (94.75%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    bool helper(TreeNode* root, long long low, long long high)
+    bool helper(TreeNode* root, long low, long high)
     {
         if(!root) return true;
         if(root->val <= low || root->val >= high)
@@ -19,6 +19,6 @@ public:
         return helper(root->left, low, root->val) && helper(root->right, root->val, high);
     }
     bool isValidBST(TreeNode* root) {
-        return helper(root, LLONG_MIN, LLONG_MAX);
+        return helper(root, LONG_MIN, LONG_MAX);
     }
 };
